backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-genai
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
reportlab==4.0.7
markdown==3.5.1